refactor(admin): cải thiện UI/UX và logic xử lý trên toàn bộ trang quản lý

+++ backend/api/v1/admin/controllers/promotion.controller.js
@@ -333,6 +333,11 @@ export const changeMulti = async (req, res) => {
         break;
 
       case "delete-all":
+        if (!req.user.roleId.permissions.includes("promotions_delete")) {
+          return res.status(403).json({
+            message: "Bạn không có quyền xóa khuyến mãi",
+          });
+        }
         await Promotion.deleteMany({
           _id: { $in: ids },
           deleted: true,
